Added a lot of styling to lists, navbar, menu, cards. App starting to look good
