+++ src/components/television/TelevisionPage.js
@@ -4,17 +4,23 @@ import { Header, Icon } from 'semantic-ui-react';
 import FilterTelevision from '../../containers/television/FilterTelevision';
 import VisibleTelevisionList from '../../containers/television/VisibleTelevisionList';
 
+const filterPadding = {
+  paddingLeft: '50px'
+};
+
 const MoviePage = props => {
   const pageHeader = props.loggedIn
     ? `Television You've Watched:`
     : `Television Others Have Watched:`;
   return (
     <div>
-      <Header as="h2">
-        <Icon name="tv" />
-        <Header.Content>{pageHeader}</Header.Content>
-      </Header>
-      <FilterTelevision />
+      <div style={filterPadding}>
+        <Header as="h2">
+          <Icon name="tv" />
+          <Header.Content>{pageHeader}</Header.Content>
+        </Header>
+        <FilterTelevision />
+      </div>
       <br />
       <VisibleTelevisionList />
     </div>
